requirements.txt
pytelegrambotapi
psycopg2
python-dotenv
